slog: Secure logging functionality moved to separature module
The secure logging functionality providing forward integrity and confidentiality for logs
is no longer part of the standard syslog-ng module cryptofuncs. It has been moved to its
own modules called secure-logging in order to allow for easier test and integration.
