+++ modules/CMakeLists.txt
@@ -39,6 +39,7 @@ add_subdirectory(python)
 add_subdirectory(redis)
 add_subdirectory(riemann)
 add_subdirectory(afsnmp)
+add_subdirectory(secure-logging)
 add_subdirectory(stardate)
 add_subdirectory(syslogformat)
 add_subdirectory(systemd-journal)
@@ -46,4 +47,3 @@ add_subdirectory(system-source)
 add_subdirectory(tagsparser)
 add_subdirectory(timestamp)
 add_subdirectory(xml)
-
