+++ modules/Makefile.am
@@ -40,6 +40,7 @@ include modules/python/Makefile.am
 include modules/redis/Makefile.am
 include modules/riemann/Makefile.am
 include modules/afsnmp/Makefile.am
+include modules/secure-logging/Makefile.am
 include modules/stardate/Makefile.am
 include modules/syslogformat/Makefile.am
 include modules/systemd-journal/Makefile.am
@@ -59,7 +60,7 @@ SYSLOG_NG_MODULES	=	\
 	mod-python mod-java mod-java-modules mod-kvformat mod-date \
 	mod-native mod-cef mod-add-contextual-data mod-diskq mod-getent \
 	mod-map-value-pairs mod-tags-parser mod-xml \
-	mod-appmodel mod-openbsd mod-snmp
+	mod-appmodel mod-openbsd mod-snmp mod-secure-logging
 
 modules modules/: ${SYSLOG_NG_MODULES}
 
@@ -72,7 +73,7 @@ modules_test_subdirs	=	\
 	modules_csvparser modules_dbparser modules_basicfuncs \
 	modules_cryptofuncs modules_geoip2 modules_afstomp \
 	modules_graphite modules_riemann modules_python \
-	modules_systemd_journal modules_kvformat modules_date \
+	modules_secure-logging modules_systemd_journal modules_kvformat modules_date \
 	modules_cef modules_diskq modules-add-contextual-data modules_getent \
 	modules_map-value-pairs modules_tagsparser modules_xml modules_appmodel
