+++ modules/secure-logging/slogkey/slogkey.c
