+++ modules/secure-logging/slogverify/slogverify.c
