+++ modules/secure-logging/tests/CMakeLists.txt
@@ -1,0 +1,1 @@
+add_unit_test(LIBTEST CRITERION TARGET test_secure_logging DEPENDS secure_logging)
